+++ packages/polymarket/src/errors.ts
@@ -7,3 +7,13 @@ export class DiscoveryError extends Error {
     this.code = code;
   }
 }
+
+export class MarketDataError extends Error {
+  readonly code: string;
+
+  constructor(message: string, code: string, options?: { cause?: unknown }) {
+    super(message, options);
+    this.name = "MarketDataError";
+    this.code = code;
+  }
+}